+++ spec/rubocop/cop/style/redundant_parentheses_spec.rb
@@ -98,6 +98,8 @@ describe RuboCop::Cop::Style::RedundantParentheses do
   it_behaves_like 'redundant', '(+x)', '+x', 'an unary operation'
   it_behaves_like 'plausible', '(!x).y'
 
+  it_behaves_like 'redundant', '[(1)]', '[1]', 'a literal', '(1)'
+
   it 'accepts parentheses around a method call with unparenthesized ' \
      'arguments' do
     inspect_source(cop, '(a 1, 2) && (1 + 1)')