+++ lib/rubocop/cop/style/space_around_operators.rb
@@ -41,7 +41,7 @@ module RuboCop
             on_special_asgn(node)
           elsif !node.unary_operation? && !called_with_dot?(node)
             op = node.method_name
-            if op != :[] && op != :! && operator?(op)
+            if op != :[] && op != :! && op != :[]= && operator?(op)
               _, _, right, = *node
               check_operator(node.loc.selector, right.source_range)
             end