+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `Style/RedundantParentheses` doesn't flag `method (:arg) {}`, since removing the parentheses would change the meaning of the expression. ([@alexdowad][])
 * `Performance/Detect` doesn't flag code where `first` or `last` takes an argument, as it cannot be transformed to equivalent code using `detect`. ([@alexdowad][])
 * `Style/SpaceAroundOperators` ignores aref assignments. ([@alexdowad][])
+* `Style/RescueModifier` indents code correctly when auto-correcting. ([@alexdowad][])
 
 ### Changes